test: Add autowired test code
Description:
The @Autowired annotation has required=true by default. If you try to inject a dependency that is not registered as a bean while required is set to true, an error will occur.
To avoid this, you can set required=false, use the @Nullable annotation, or use Optional to handle the absence of a bean gracefully.